Move file content utilities into types/files.ts and FilePreviewSheet
Consolidate MIME type handling in types/files.ts: content type checkers
(isOfficeViewerCompatible, isPdfContentType, isMarkdownContentType),
content processing (processFileContent, csvToMarkdownTable), and types
(ProcessedContent). Move preview-specific logic (getFilePreviewConfig,
FilePreviewConfig, FilePreviewCategory) into FilePreviewSheet where
it's used. Delete lib/file_content_utils.ts.
