+++ front/pages/api/w/[wId]/files/[fileId]/office-viewer-url.ts
@@ -2,11 +2,11 @@ import type { NextApiRequest, NextApiResponse } from "next";
 
 import { withSessionAuthenticationForWorkspace } from "@app/lib/api/auth_wrappers";
 import type { Authenticator } from "@app/lib/auth";
-import { isOfficeViewerCompatible } from "@app/lib/file_content_utils";
 import { FileResource } from "@app/lib/resources/file_resource";
 import { SpaceResource } from "@app/lib/resources/space_resource";
 import { apiError } from "@app/logger/withlogging";
 import type { WithAPIErrorResponse } from "@app/types";
+import { isOfficeViewerCompatible } from "@app/types/files";
 
 export interface OfficeViewerUrlResponseBody {
   signedUrl: string;
